Add files via upload
This is the 'Heart-Disease-Prediction' project. The dataset is from 'kaggle'. It predicts whether the specific person has Hear-disease or not based on his/her results of various test that've been conducted by the clinic. It contains features like cholesterol, age, sex, thalach and more. Based on this data, the prediction is done.